LayoutTests/platform/blackberry/media/video-controls-enlarged-fullscreen-nometa-expected.txt

Test that media controls are rendered at the enlarged size in fullscreen mode.

EVENT(canplaythrough)
The actual controls height matches the expected controls height. OK
END OF TEST
Enter FullScreen
